feature(data): add fiche sp fetcher
