+++ packages/code-du-travail-data/dataset/fiches_service_public/format.js
@@ -1,4 +1,4 @@
-const parseReference = require("./parseReference");
+const parseReference = require("./parse-reference");
 const getChild = (element, name) => element.$.find(el => el.name === name);
 
 // Beware, this one is recursive
