+++ packages/code-du-travail-data/dataset/fiches_service_public/index.js
@@ -8,34 +8,35 @@ const format = require("./format");
 
 const read = path => fs.readFileSync(path).toString();
 
-const getFiches = fichePath => fs
-  .readdirSync(path.join(__dirname, fichePath))
-  .filter(file => file.match(/F[0-9]+/))
-  .map(file =>read(`${fichePath}/${file}`))
+const getFiches = fichePath =>
+  fs
+    .readdirSync(path.join(__dirname, fichePath))
+    .filter(file => file.match(/F[0-9]+/))
+    .map(file => read(`${fichePath}/${file}`));
 
 const fiches = [].concat(
-  getFiches("data/vosdroits-particuliers"),
-  getFiches("data/vosdroits-professionnels"),
-  getFiches("data/vosdroits-associations")
+  getFiches("data/particuliers"),
+  getFiches("data/professionnels"),
+  getFiches("data/associations")
 );
 
-const parsedFiches = fiches.map(fiche => xmlStringToJsObject(fiche, {
-  alwaysArray: true,
-  ignoreDeclaration: true,
-  ignoreDoctype:  true,
-  ignoreInstruction: true,
-  elementsKey: "$",
-  attributesKey: "_",
-  textKey: "$"
-}))
+const parsedFiches = fiches.map(fiche =>
+  xmlStringToJsObject(fiche, {
+    alwaysArray: true,
+    ignoreDeclaration: true,
+    ignoreDoctype: true,
+    ignoreInstruction: true,
+    elementsKey: "$",
+    attributesKey: "_",
+    textKey: "$"
+  })
+);
 
-const uniqFiches = uniqBy(parsedFiches, fiche => fiche.$[0]._.ID)
+const uniqFiches = uniqBy(parsedFiches, fiche => fiche.$[0]._.ID);
 const filteredFiches = filter(uniqFiches);
 
-const formatedFiches = filteredFiches
-  .map(format)
-  .filter(Boolean);
+const formatedFiches = filteredFiches.map(format).filter(Boolean);
 
 if (module === require.main) {
-  console.log(JSON.stringify(formatedFiches, null, 2))
+  console.log(JSON.stringify(formatedFiches, null, 2));
 }
