--- packages/code-du-travail-data/dataset/fiches_service_public/updateTheme.js
@@ -1,20 +1,0 @@
-const getThemeMapping = require("./cdtn-theme-sp.js");
-
-async function updateTheme(fiches) {
-  const themeMapping = await getThemeMapping()
-  fiches.forEach(fiche => {
-    fiche.themeCdtn = themeMapping[fiche.id] || [];
-  });
-}
-
-async function main() {
-  const fiches = require('./fiches-sp-travail.json');
-  await updateTheme(fiches);
-  console.log(JSON.stringify(fiches, null, 2))
-}
-
-if (module === require.main) {
-  main()
-}
-
-module.exports = updateTheme
